Implement users search by email. Add API method to fetch users by wallet. Refactor wallet actions controller.
